Move dynamical coring to md submodule.
