+++ src/msmhelper/md/__init__.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 """  --8<-- "README.md" """
 
+from .corrections import dynamical_coring
 from .timescales import (
     estimate_waiting_times,
     estimate_wt,
