+++ test/test_md_corrections.py
@@ -2,17 +2,15 @@
 """Tests for the dynamical correction module.
 
 BSD 3-Clause License
-Copyright (c) 2019-2022, Daniel Nagel
+Copyright (c) 2019-2023, Daniel Nagel
 All rights reserved.
-
-Author: Daniel Nagel
 
 """
 import numba
 import numpy as np
 import pytest
 
-from msmhelper import dyncor
+from msmhelper.md import corrections
 from msmhelper.statetraj import LumpedStateTraj, StateTraj
 
 
@@ -30,7 +28,7 @@ def test__remains_in_core(idx, traj, lagtime, iterative, result):
     if not numba.config.DISABLE_JIT:
         traj = numba.typed.List(traj)
 
-    assert dyncor._remains_in_core(
+    assert corrections._remains_in_core(
         idx=idx,
         traj=traj,
         lagtime=lagtime,
@@ -49,7 +47,7 @@ def test__find_first_core(traj, lagtime, result):
     if not numba.config.DISABLE_JIT:
         traj = numba.typed.List(traj)
 
-    assert dyncor._find_first_core(traj, lagtime) == result
+    assert corrections._find_first_core(traj, lagtime) == result
 
 
 @pytest.mark.parametrize('traj, lagtime, iterative, result, error', [
@@ -86,7 +84,7 @@ def test__find_first_core(traj, lagtime, result):
         6,
         True,
         None,
-        dyncor.LagtimeError,
+        corrections.LagtimeError,
     ),
 ])
 def test__dynamical_coring_single_traj(
@@ -98,7 +96,7 @@ def test__dynamical_coring_single_traj(
         traj = numba.typed.List(traj)
 
     if error is None:
-        cored_traj = dyncor._dynamical_coring_single_traj(
+        cored_traj = corrections._dynamical_coring_single_traj(
             traj=traj,
             lagtime=lagtime,
             iterative=iterative,
@@ -106,7 +104,7 @@ def test__dynamical_coring_single_traj(
         np.testing.assert_array_almost_equal(cored_traj, result)
     else:
         with pytest.raises(error):
-            dyncor._dynamical_coring_single_traj(
+            corrections._dynamical_coring_single_traj(
                 traj=traj,
                 lagtime=lagtime,
                 iterative=iterative,
@@ -170,7 +168,7 @@ def test__dynamical_coring_single_traj(
 def test_dynamical_coring(trajs, lagtime, kwargs, result, error):
     """Test dynamical coring."""
     if error is None:
-        cored_trajs = dyncor.dynamical_coring(
+        cored_trajs = corrections.dynamical_coring(
             trajs=trajs, lagtime=lagtime, **kwargs,
         )
         assert len(cored_trajs) == len(result)
@@ -178,4 +176,4 @@ def test_dynamical_coring(trajs, lagtime, kwargs, result, error):
             np.testing.assert_array_almost_equal(cored_traj, traj)
     else:
         with pytest.raises(error):
-            dyncor.dynamical_coring(trajs=trajs, lagtime=lagtime, **kwargs)
+            corrections.dynamical_coring(trajs=trajs, lagtime=lagtime, **kwargs)
